arm64 assembly function union2by2.
